backend/requirements.txt
# FastAPI & Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiomysql>=0.2.0
cryptography>=42.0.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Redis & Queue
redis[hiredis]>=5.0.0
arq>=0.25.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# LLM & Search
openai>=1.12.0
tavily>=0.3.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# MemOS Memory System
MemoryOS>=2.0.0

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utils
orjson>=3.9.0
structlog>=24.1.0

# Test
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.26.0
asyncmy>=0.2.9
email-validator>=2.1.0
bcrypt>=4.0.0,<5.0.0
